+++ indexer/packages/postgres/src/db/migrations/migration_files/20240830165511_create_affiliate_info_table.ts
@@ -3,13 +3,13 @@ import * as Knex from 'knex';
 export async function up(knex: Knex): Promise<void> {
   return knex.schema.createTable('affiliate_info', (table) => {
     table.string('address').primary().notNullable();
-    table.integer('affiliateEarnings').notNullable();
+    table.decimal('affiliateEarnings').notNullable();
     table.integer('referredMakerTrades').notNullable();
     table.integer('referredTakerTrades').notNullable();
-    table.integer('totalReferredFees').notNullable();
+    table.decimal('totalReferredFees').notNullable();
     table.integer('totalReferredUsers').notNullable();
-    table.integer('referredNetProtocolEarnings').notNullable();
-    table.integer('firstReferralBlockHeight').notNullable();
+    table.decimal('referredNetProtocolEarnings').notNullable();
+    table.bigInteger('firstReferralBlockHeight').notNullable();
   });
 }
